requirements.txt
telethon
cryptg
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
requests
